+++ packages/frontend/src/hooks/transaction/useConvertTransactionToTransactionListingItem.ts
@@ -1,0 +1,61 @@
+import { useCallback } from 'react';
+
+import { getTransactionType } from '../../utils/transaction/getTransactionType';
+
+import {
+  TransactionCategoryMappingDto,
+  TransactionDto,
+} from '$api/generated/financerApi';
+import {
+  TransactionListingItemProps,
+  TransactionType,
+} from '$blocks/transaction-listing/transaction-listing.item';
+import { formatCurrency } from '$utils/formatCurrency';
+import { formatDate } from '$utils/formatDate';
+
+type TransactionDtoForConvert = Omit<
+  TransactionDto,
+  'fromAccount' | 'toAccount'
+> & {
+  fromAccount?: string;
+  toAccount?: string;
+};
+
+const mapTransactionTypeToUrlPrefix: {
+  [key in TransactionType]: 'incomes' | 'expenses' | 'transfers';
+} = {
+  income: 'incomes',
+  expense: 'expenses',
+  transfer: 'transfers',
+};
+
+export const useConvertTransactionToTransactionListingItem = () => {
+  return useCallback(
+    (
+      transaction: TransactionDtoForConvert,
+      getCategoryName: (id: string) => string | undefined
+    ): TransactionListingItemProps => {
+      const transactionType = getTransactionType(
+        transaction.toAccount,
+        transaction.fromAccount
+      );
+
+      return {
+        transactionCategories: (
+          transaction.categories as unknown as TransactionCategoryMappingDto[]
+        )
+          .map(({ category_id }) =>
+            getCategoryName(category_id as unknown as string)
+          )
+          .join(', '),
+        transactionAmount: formatCurrency(transaction.amount),
+        date: formatDate(new Date(transaction.date)),
+        label: transaction.description,
+        link: `/statistics/${mapTransactionTypeToUrlPrefix[transactionType]}/${transaction._id}`,
+        transactionType,
+        id: transaction._id,
+      };
+    },
+    []
+  );
+};
